requirements.txt
django==3.2.9
djangorestframework==3.12.4
drf_yasg==1.20.0
psycopg2==2.9.2
pytest==6.2.5
pytest-django==4.4.0
requests==2.26.0